requirements.txt
app
tweepy==3.6.0
speedtest-cli==2.0.0